requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
decorator==5.1.1
entrypoints==0.4
ffmpeg-python==0.2.0
filelock==3.12.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
GoogleBard==0.0.7
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
libretranslatepy==2.1.1
llvmlite==0.40.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numba==0.57.0
numpy==1.24.3
openai-whisper==20230314
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
playsound==1.2.2
prompt-toolkit==3.0.38
protobuf==3.20.3
pyarrow==12.0.0
PyAudio==0.2.13
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypiwin32==223
pyrsistent==0.19.3
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3
pywin32==306
regex==2023.5.5
requests==2.31.0
rfc3986==1.5.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SpeechRecognition==3.10.0
streamlit==1.22.0
sympy==1.12
tenacity==8.2.2
textblob==0.17.1
tiktoken==0.3.1
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
translate==3.6.1
typing_extensions==4.6.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.15.0
